+++ dst/app/routes/payment/any.js
@@ -32,7 +32,7 @@ const anyPaymentRouter = express_1.Router();
  * 汎用決済承認
  */
 // tslint:disable-next-line:use-default-type-parameter
-anyPaymentRouter.post('/authorize', permitScopes_1.default([iam_1.Permission.User, 'pos']), rateLimit_1.default, ...[
+anyPaymentRouter.post('/authorize', permitScopes_1.default([iam_1.Permission.User, 'pos', 'payment.any.write']), rateLimit_1.default, ...[
     express_validator_1.body('object')
         .not()
         .isEmpty(),
@@ -95,7 +95,7 @@ anyPaymentRouter.post('/authorize', permitScopes_1.default([iam_1.Permission.Use
 /**
  * 汎用決済承認取消
  */
-anyPaymentRouter.put('/authorize/:actionId/void', permitScopes_1.default([iam_1.Permission.User]), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+anyPaymentRouter.put('/authorize/:actionId/void', permitScopes_1.default([iam_1.Permission.User, 'payment.any.write']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     yield rateLimit4transactionInProgress_1.default({
         typeOf: req.body.purpose.typeOf,
         id: req.body.purpose.id